PLAT-966: node pools should start at the minimum count
When deploying on AKS, multiple nodes may be created in the same zone.
The scheduler will use all the available nodes. However, the autoscaler
may attempt to scale down. This can happen during a deploy. In some
unlucky situations, it can take minutes to re-attach disks when
switching nodes. Helm has a tendency to fail if it happens to be some
dependency like postgres or mongo.
To prevent such issues, we will do the opposite and scale up if needed.